backend/pyproject.toml
[tool.poetry]
name = "portal-swissit"
version = "0.1.0"
description = "SwissIT Design client portal"
authors = ["Lucien Rey <lucien.rey@swissit-design.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
asgiref = "^3.8.1"
django-cors-headers = "^4.4.0"
pyjwt = "^2.8.0"
sqlparse = "^0.5.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
whitenoise = "^6.7.0"
django-ninja-jwt = "^5.3.2"
django-ninja-extra = "^0.21.1"
django-ninja-passwordreset = "^1.2.5"
django-extensions = "^3.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
